+++ demos/default/pages/getStaticProps/withFallbackAndMiddleware/_middleware.js
@@ -1,0 +1,7 @@
+import { NextResponse } from 'next/server'
+
+export function middleware() {
+  const res = NextResponse.next()
+  res.headers.set('x-middleware-date', new Date().toISOString())
+  return res
+}